requirements.txt
pyqt5>=5.15.4
opencv-python>=4.5.1.48
torch>=1.8.1+cu111
numpy>=1.21.0
pandas>=1.1.5
torchvision>=0.9.1+cu111
pyyaml>=5.4.1
matplotlib>=3.2.2
thop>=0.0.31-2005241907
requests~=2.25.1
cython>=0.29.23
dlib>=19.22.0
pillow>=8.2.0
future>=0.18.2
tqdm>=4.57.0
tensorboardx>=2.2
easydict>=1.9
qdarkstyle>=3.0.2